Refactored how we do the build_id parameter in light of folks starting to use 'all' command
It was checking for build_id parameter but it can be
blank and needs to be if you want to build the trunk.

So moved the parameter check out of init and initialize
the build id to the date if it is missing when we go to
do the packaging.