Spliting contrail-openstack-vrouter to contrail-vrouter-common and orchestrator specific packages.
In contrail only deployment with Vcenter as Orchestrator we no need to install the following packages
in the compute nodes,nova-compute, python-novaclient, contrail-nova-vif. However when
contrail-openstack-vrouter is installed, it also installs opensatack specific packagesas the
contrail-openstack-vrouter is dependent on them.

Fix is to come up with contrail-vrouter-common package, which will depend on the common dependent packages
and make contrail-openstack-vrouter depend on contrail-vrouter-common and nova(openstack) related packages.
For any new Orchestrator like vcenter, we need to come up with contrail-vcenter-vrouter, if it is dependent
on contrail-vrouter-common and some new packages(vcenter/esxi specific), else we could use the
contrail-vrouter-common to install compute node.

Change-Id: I8e23a63d43d858c0e94e9101f91e8bbbd755e09a
Closes-Bug: 1392925