Make nodemgr for vrouter node poll for process status.
Currently we have seen scenarios where vrouter nodemgr
comes up later than contrail-vrouter-agent and does not
report that in the nodestatus uve.

In this change we call supervisorctl to check process
status every 60 seconds and update any process status
which is not in our list.

We have not made the change applicable for other nodetypes
as nodemgrs for other nodetypes do not run as root.

Change-Id: Icdb65f5c9cdfcb9804248730bf398696a1660c63
Closes-Bug: 1415266