nodemgr exits when discovery is not configured for all services
Issue:
If discovery is not configured for a service then nodemgr should not exit but log error and continue.
Discovery Port should also be read from conf file if discovery server is read from conf file.

Fix:
catch an exception if discovery is not configured a service and log an error.
Add code to read discovery port from conf file.

Change-Id: Ic25c6457170143474d0e1352627a49d7ddf2695f
Closes-Bug: 1422157