Support for vhost-user nic for non-MSI guests
This provides an updated qemu packages with the patch allowing for
vhost-user nic to be used with non-MSI guests. This is necessary for vMX
to work correctly on DPDK vRouter compute nodes.

Patch backported from upstream qemu:
http://git.qemu.org/?p=qemu.git;a=commitdiff;h=24f938a682d934b133863eb421aac33592f7a09e

New qemu*.deb and dpdk-depends-packages_0.4-0contrail1_amd64.deb packages
has to be put under /cs-shared

Change-Id: I3f82401ac9178585979c9e7a2bf3354faa4e5223
Closes-Bug: #1396396