DPDK: remove bond interface on start
When vRouter uses a VLAN, it tries to create a VLAN forwarding interface
using KNI. By default it uses the original physical interface name.

In case of normal ethernet interfaces (i.e. eth0), the interface
disappears from the Linux after the unbind.

In case of bond interfaces (i.e. bond0), the slaves of the bond disappear
from the Linux, but the bond interface is still there.

So when vRouter uses VLAN and bond at the same time, it fails to create
a KNI with the original bond interface name.

ifdown removed the bond interface form the Linux and fixes the issue.

+ minor fixes

Change-Id: I15ba3404a92e1e7dd70773ba4e697d20e944a31d
Closes-bug: 1484021