DPDK: vrouter fails to start after supervisor-vrouter restart
In case of vlan setup, after restarting supervisor-vrouter service (service supervisor-vrouter restart)
does not come-up

The main problem is that, after rebinding interface to kernel driver (from DPDK driver)
a Linux kernel is not fast enough to create vlan interface.

I just create a loop which tries checks if interface exists.
The file which is used to check existence of interface is "/sys/class/net/${INTERFACE}"
Maximum timeout is currently set to 20 seconds.


Change-Id: Ifc620412d5fb9e0fe1e374548b7cf82d451625b7
Closes-bug: #1508729