DPDK: fix NUMA node ID to EAL for bond interface
If /sys/class/{dev}/device/numa_node returns -1 for bond interface
and it's passed to EAL via vRouter's --vdev socket_id parameter,
it gets interpreted as 255. This for sure is greater than the
number of NUMA nodes in system, therefore vRouter will crash
on startup.

If we get -1, pretend it's 0, so DPDK EAL does not complain.

Change-Id: I028fb5842335df93eaa824a2fdc621218ce0a95e
Closes-bug: #1531596