Fix for supervisor-vrouter restart
In case of bonding, supervisor-vrouter restart was causing bonding configuration in the NIC to be reset.
When the dpdk-vrouter comes up again, bonding wont work since NIC is not initialized with bonding.
Fix is to re-initialized networking when vrouter is stopped so that bonding configuration is re-applied to the NIC.

Change-Id: I7099947ef30492d6906641f41c927d3ec1cbe0a7
Closes-Bug: 1526133