DPDK: Fix DPDK vRouter start/stop scripts
The insert_vrouter() function from vrouter-functions.sh, when in DPDK mode
should not be called every time vhost0 interface goes up, but when the vRouter
process starts. Thus, we should have our own, phony, version of if-vhost0
scripts that does nothing. We need it as other scripts, which modifies
/etc/network/interfaces place explicit: 'pre-up /opt/contrail/bin/if-vhost0'
line to the vhost0 interface.

Also, there is no need for the vrouter_dpdk_start() function to call:
'service supervisor-vrouter start' explicitly since it is called by the
post-start script of the supervisor-vrouter service itself.

This commit fixes the issue that caused the ifup scripts to always bind the
network interface to the DPDK (which removes the iface from being visible in
the OS) even though, the DPDK vRouter was not started.

Change-Id: I945484462ab8cef9eb077b7eab0692c7e32f971a
Closes-Bug: #1543637