Work around race between vRouter and network interfaces start
There were observed cases when the vRouter had started before the bond0
interface was created, which led to an error as DPDK could not bind to the
non-existing interface.  However, this code is only a TEMPORARY HACK. In
overall the entire file is nothing more than a set of ugly or very ugly hacks
and should be rewritten as soon as possible. All functions depend on each
other, there is no way to add any functionality other than introducing yet
another hack before it is refactored.

Also, it removes the loop that waits for the existence of a device in
_dpdk_conf_read() function as it was placed in the wrong place. The
_dpdk_conf_read() is used also when vRouter is already started. In this case,
there is no interface as it has been taken by the DPDK and the script always
added at least 20s of delay, enlarging the start/stop time of
supervisor-vrouter service.

Change-Id: I24e9401a84cd641f71a088ad568c6e6432c5e80d
Closes-Bug: #1544213