Get rid of dpdk-depends-packages in Liberty/Mitaka
For Liberty and Mitaka we have to patch the libvirt anyway since it is
required for multiqueue support on both DPDK and kernel vRouter.

The patches in qemu for DPDK vRouter consists only of few lines and does not
affect the kernel vRouter behaviour so the packages can be shared between
kernel and DPDK vRouter setups.

That way we can eliminate the necessity of the dpdk-depends-packages, what
this commit is doing.

Change-Id: Ic96364b1125f0969ead5f1a7aaf3e17659f78c29
Closes-Bug: #1595561