Adding python-memcache to contrail-openstack-config dependency of juno
The issue is seen in juno after updating the keystonemiddleware package
version used in juno, fixing it in control.in for juno.

The latest keystonemiddleware contrail-config uses keystonemiddleware to authenticate with keystone.
keystonemiddleware has a dependency on memcache, but the dependency is
not defined in the upstream keystonemiddleware package.
Adding the dependency in contrail-openstack-config as the config node
needs memcache installed while using keystonemiddleware.

Change-Id: Ie228f37358df1c699859b8b3cb6b26a62893acf2
Closes-Bug: #1602618