Closes-Bug: #1647157 - Cliff client is now default client in mainline - Old client shouldn't be installed
This bug will only be seen in master due to this commit:
https://github.com/Juniper/contrail-packaging/commit/a38e9c02048a9f4c30efdc2400064bd317bae182

The comment here mentioned that cliff client is default in mainline:
https://bugs.launchpad.net/juniperopenstack/r3.0.2.x/+bug/1635035/comments/22

But the package is still being installed as part of --all
This will be removed as part of this commit
This check-in also brings in couple of python scripts and some sample JSON files that were previously being brought in
by the SM client package.
These files are needed for SM lite install (preconfig.py, testbed_parser.py)

After this check-in, Cliff client and old client CANNOT be installed on same machine.

Closes-Bug: #1635035 - Cliff client made default on mainline only

Will fix mainline such that sm-client is not installed and only sm-cliff-client is

Patch 2: Fixed SM src path for client files

Change-Id: Ia9730d040a9b82ee175de4d3db628e6e4e2bcd53