Disable opscenterd in postinstall of contrail-openstack-webui
Opscenter can be used to access cassandra without authentication
since we do not enable JMX authentication. Till we enable it, we
will disable opscenter from running by default but still install
it. This will be done in postinstall script of contrail-openstack-webui
since opscenter is installed as part of that package.

Change-Id: I8a738365ac91c4c90f98d36ea84d6d0f6af33729
Closes-Bug: #1688637