Changing the redis port for SMLite UI from 6379 to 6380
redis-server listens on both controller ip and localhost in container.
In controller container, web-ui config file gets redis_ip changed to
connect to controller ip.

Now in host we run redis-server on 127.0.0.1:6379,

so while in controller docker, webui provisioning restart of redis-server,
fails, as 6379 is already binded with, so changing the redis port
for SMLite from 6379 to 6380.

Change-Id: Iea83cec3732633bba0e7807fdd50ca6f97db1f3b
Closes-Bug: #1704215